Moved code to the new journey folders
